split out ability & controllable
