+++ test/ability.ts
@@ -1,10 +1,10 @@
-import {expect, fixture, html} from '@open-wc/testing'
-import {restore, fake} from 'sinon'
 import type {CustomElement} from '../src/custom-element.js'
-import {createAbility, attachShadowCallback, attachInternalsCallback} from '../src/ability.js'
+import {expect, fixture, html} from '@open-wc/testing'
+import {restore} from 'sinon'
+import {createAbility} from '../src/ability.js'
 
 describe('ability', () => {
-  let calls: string[] = []
+  const calls: string[] = []
   const fakeable = createAbility(
     Class =>
       class extends Class {
@@ -90,190 +90,5 @@ describe('ability', () => {
     expect(calls).to.eql(['fakeable connectedCallback'])
     instance.connectedCallback!()
     expect(calls).to.eql(['fakeable connectedCallback', 'fakeable connectedCallback'])
-  })
-
-  describe('subclass behaviour', () => {
-    const CoreTest = otherfakeable(fakeable(Element))
-    customElements.define('core-test', CoreTest)
-
-    let instance: CustomElement & typeof CoreTest
-    beforeEach(async () => {
-      instance = await fixture(html`<core-test />`)
-    })
-
-    it('applies keys from delegate onto subclass upon instantiation', () => {
-      expect(instance).to.have.property('foo')
-      expect((instance as unknown as Record<string, () => void>).foo()).to.equal('foo!')
-      expect(instance).to.have.property('bar')
-      expect((instance as unknown as Record<string, () => void>).bar()).to.equal('bar!')
-    })
-
-    for (const method of ['connectedCallback', 'disconnectedCallback', 'adoptedCallback', 'attributeChangedCallback']) {
-      it(`delegates to other ${method}s before class ${method}`, () => {
-        calls = []
-        ;(instance as unknown as Record<string, () => void>)[method]()
-        expect(calls).to.eql([`otherfakeable ${method}`, `fakeable ${method}`])
-      })
-    }
-  })
-
-  describe('ability extension behaviour', () => {
-    describe('attachShadowCallback', () => {
-      let attachShadowFake: (shadow: ShadowRoot) => void
-      let shadow: ShadowRoot | null
-      beforeEach(() => {
-        shadow = null
-        attachShadowFake = fake()
-      })
-
-      const declarable = createAbility(
-        Class =>
-          class extends Class {
-            [attachShadowCallback](...args: [ShadowRoot]) {
-              super[attachShadowCallback]!(...args)
-              return attachShadowFake.apply(this, args)
-            }
-          }
-      )
-      customElements.define(
-        'declarative-shadow-ability',
-        declarable(
-          class extends HTMLElement {
-            constructor() {
-              super()
-              // Declarative shadows run before constructor() is available, but
-              // abilities run after element constructor
-              shadow = HTMLElement.prototype.attachShadow.call(this, {mode: 'closed'})
-            }
-          }
-        )
-      )
-      customElements.define(
-        'closed-shadow-ability',
-        declarable(
-          class extends HTMLElement {
-            constructor() {
-              super()
-              shadow = this.attachShadow({mode: 'closed'})
-            }
-          }
-        )
-      )
-      customElements.define(
-        'connected-shadow-ability',
-        declarable(
-          class extends HTMLElement {
-            connectedCallback() {
-              shadow = this.attachShadow({mode: 'closed'})
-            }
-          }
-        )
-      )
-      customElements.define('manual-shadow-ability', declarable(class extends HTMLElement {}))
-
-      customElements.define(
-        'disallowed-shadow-ability',
-        declarable(
-          class extends HTMLElement {
-            static disabledFeatures = ['shadow']
-          }
-        )
-      )
-
-      it('is called with shadowRoot of declarative ShadowDOM', async () => {
-        const instance = await fixture(html`<declarative-shadow-ability></declarative-shadow-ability>`)
-        expect(shadow).to.exist.and.be.instanceof(ShadowRoot)
-        expect(attachShadowFake).to.be.calledOnce.calledOn(instance).and.calledWithExactly(shadow)
-      })
-
-      it('is called with shadowRoot from attachShadow call', async () => {
-        const instance = await fixture(html`<manual-shadow-ability></manual-shadow-ability>`)
-        shadow = instance.attachShadow({mode: 'closed'})
-        expect(shadow).to.exist.and.be.instanceof(ShadowRoot)
-        expect(attachShadowFake).to.be.calledOnce.calledOn(instance).and.calledWithExactly(shadow)
-      })
-
-      it('is called with shadowRoot from attachInternals call', async () => {
-        const instance = await fixture(html`<closed-shadow-ability></closed-shadow-ability>`)
-        expect(shadow).to.exist.and.be.instanceof(ShadowRoot)
-        expect(attachShadowFake).to.be.calledOnce.calledOn(instance).and.calledWithExactly(shadow)
-      })
-
-      it('is called with shadowRoot from connectedCallback', async () => {
-        const instance = await fixture(html`<connected-shadow-ability></connected-shadow-ability>`)
-        expect(shadow).to.exist.and.be.instanceof(ShadowRoot)
-        expect(attachShadowFake).to.be.calledOnce.calledOn(instance).and.calledWithExactly(shadow)
-      })
-
-      it('does not error if shadowdom is disabled', async () => {
-        await fixture(html`<disabled-shadow-ability></disabled-shadow-ability>`)
-        expect(attachShadowFake).to.be.have.callCount(0)
-      })
-    })
-
-    describe('attachInternalsCallback', () => {
-      let attachInternalsFake: (internals: ElementInternals) => void
-      let internals: ElementInternals | null
-      beforeEach(() => {
-        internals = null
-        attachInternalsFake = fake()
-      })
-
-      const internable = createAbility(
-        Class =>
-          class extends Class {
-            [attachInternalsCallback](...args: [ElementInternals]) {
-              super[attachInternalsCallback]!(...args)
-              return attachInternalsFake.apply(this, args)
-            }
-          }
-      )
-      customElements.define(
-        'internals-ability',
-        internable(
-          class extends HTMLElement {
-            constructor() {
-              super()
-              internals = this.attachInternals()
-            }
-          }
-        )
-      )
-      customElements.define('manual-internals-ability', internable(class extends HTMLElement {}))
-
-      customElements.define(
-        'disallowed-internals-ability',
-        internable(
-          class extends HTMLElement {
-            static disabledFeatures = ['internals']
-          }
-        )
-      )
-
-      it('is called on constructor', async () => {
-        const instance = await fixture(html`<manual-internals-ability></manual-internals-ability>`)
-        expect(attachInternalsFake).to.be.calledOnce.calledOn(instance)
-      })
-
-      it('does not prevent attachInternals being called by userland class', async () => {
-        const instance = await fixture(html`<internals-ability></internals-ability>`)
-        expect(internals).to.exist.and.be.instanceof(ElementInternals)
-        expect(attachInternalsFake).to.be.calledOnce.calledOn(instance).and.calledWithExactly(internals)
-      })
-
-      it('errors if userland calls attachInternals more than once', async () => {
-        const instance = await fixture<CustomElement>(html`<manual-internals-ability></manual-internals-ability>`)
-        internals = instance.attachInternals()
-        expect(internals).to.exist.and.be.instanceof(ElementInternals)
-        expect(attachInternalsFake).to.be.calledOnce.calledOn(instance).and.calledWithExactly(internals)
-
-        expect(() => instance.attachInternals()).to.throw(DOMException)
-      })
-
-      it('does not error if element internals are disabled', async () => {
-        await fixture(html`<disallowed-internals-ability></disallowed-internals-ability>`)
-        expect(attachInternalsFake).to.have.callCount(0)
-      })
-    })
   })
 })
